app/requirements.txt
Flask
gunicorn
python-dotenv
boto3
botocore
flask-wtf
requests
flask-paginate